Separated MyRawSocketHandler from main
